requirements.txt
mediapipe>=0.10.14
opencv-python>=4.9.0
pygame>=2.5.2
google-genai>=0.3.0
python-dotenv>=1.0.1
numpy>=1.26.0
Pillow>=10.2.0
pytest>=8.0.0
